+++ packages/office-ui-fabric-react/src/components/Persona/PersonaCoin.styles.ts
@@ -1,52 +1,9 @@
 import { PersonaInitialsColor } from './Persona.types';
 import { mergeStyles } from '../../Styling';
-
-/**
- * These colors are considered reserved colors and can only be set with overrides:
- * - Red is a color that often has a special meaning.
- * - Transparent is not intended to be used with typical initials due to accessibility issues,
- *   its primary use is for Facepile overflow buttons.
- */
-const COLOR_SWATCHES_LOOKUP: PersonaInitialsColor[] = [
-  PersonaInitialsColor.lightGreen,
-  PersonaInitialsColor.lightBlue,
-  PersonaInitialsColor.lightPink,
-  PersonaInitialsColor.green,
-  PersonaInitialsColor.darkGreen,
-  PersonaInitialsColor.pink,
-  PersonaInitialsColor.magenta,
-  PersonaInitialsColor.purple,
-  PersonaInitialsColor.black,
-  PersonaInitialsColor.teal,
-  PersonaInitialsColor.blue,
-  PersonaInitialsColor.darkBlue,
-  PersonaInitialsColor.orange,
-  PersonaInitialsColor.darkRed
-];
-
-const COLOR_SWATCHES_NUM_ENTRIES = COLOR_SWATCHES_LOOKUP.length;
-
-function getColorFromName(displayName: string | undefined): PersonaInitialsColor {
-  let color = PersonaInitialsColor.blue;
-  if (!displayName) {
-    return color;
-  }
-
-  let hashCode = 0;
-  for (let iLen: number = displayName.length - 1; iLen >= 0; iLen--) {
-    const ch: number = displayName.charCodeAt(iLen);
-    const shift: number = iLen % 8;
-    // tslint:disable-next-line:no-bitwise
-    hashCode ^= (ch << shift) + (ch >> (8 - shift));
-  }
-
-  color = COLOR_SWATCHES_LOOKUP[hashCode % COLOR_SWATCHES_NUM_ENTRIES];
-
-  return color;
-}
+import { getInitialsColorFromName } from './getInitialsColorFromName';
 
 function getInitialsColorCode(personaInitialsColor: PersonaInitialsColor | undefined, primaryText: string | undefined): string {
-  let initialsColor = personaInitialsColor !== undefined ? personaInitialsColor : getColorFromName(primaryText);
+  let initialsColor = personaInitialsColor !== undefined ? personaInitialsColor : getInitialsColorFromName(primaryText);
 
   switch (initialsColor) {
     case PersonaInitialsColor.lightBlue:
